requirements.txt
adk
google-cloud-pubsub
google-cloud-firestore
pyyaml
paho-mqtt
protobuf
google-adk
google-cloud-dialogflow-cx
grpcio-tools
google-cloud-aiplatform 
google-adk
absl-py
poetry
google-genai
pydantic
python-dotenv
pytest
pytest-asyncio